Rename *.h.inc -> *.h
