+++ experimental/split_mlir/src/iree/split_mlir/Passes.cpp
@@ -4,5 +4,5 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/split_mlir/MarkBisectPassImpl.h.inc"
-#include "iree/split_mlir/OutlineFunctionsPassImpl.h.inc"
+#include "iree/split_mlir/MarkBisectPassImpl.h"
+#include "iree/split_mlir/OutlineFunctionsPassImpl.h"
